model get Data done. Refuctoring done.
