+++ source/server/build.gradle
@@ -17,8 +17,13 @@ dependencies {
     compile 'com.fasterxml.jackson.core:jackson-databind:2.9.9.3'
     compile 'com.google.guava:guava:19.0'
     compile 'org.json:json:20180130'
+    compile 'org.javassist:javassist:3.23.1-GA'
     testCompile 'org.mockito:mockito-all:1.9.5'
     testCompile 'junit:junit:4.12'
+    implementation "javax.xml.bind:jaxb-api:2.2.11"
+    implementation "com.sun.xml.bind:jaxb-core:2.2.11"
+    implementation "com.sun.xml.bind:jaxb-impl:2.2.11"
+    implementation "javax.activation:activation:1.1.1"
 }
 
 group = 'com.registry'
