+++ source/server/src/main/java/registry/Controller.java
@@ -3,25 +3,21 @@ package registry;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.json.JSONArray;
-import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.Resource;
-import registry.dataBase.DataBase;
 import org.springframework.core.io.ByteArrayResource;
+import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import registry.menu.MenuItemEntity;
-import registry.menu.MenuItemRepository;
-import registry.model.ModelRepository;
-import registry.report.Report;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import registry.entity.menu.MenuItemRepository;
+import registry.utils.ModelUtils;
+import registry.utils.ReportUtils;
 
 import javax.sql.DataSource;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-import java.util.Map;
 
 
 @RestController
@@ -30,11 +26,11 @@ public class Controller {
     @Autowired
     private DataSource dataSource;
     @Autowired
-    private Report report;
+    private ReportUtils report;
     @Autowired
     private MenuItemRepository menuItemRepository;
     @Autowired
-    private ModelRepository modelRepository;
+    private ModelUtils modelUtils;
 
     @RequestMapping("/")
     public String index() {
@@ -54,11 +50,6 @@ public class Controller {
         return mapper.writeValueAsString(menuItemRepository.findAllRoot());
     }
 
-    @RequestMapping ("/api/directory/person")
-    public String person () {
-        return DataBase.getJsonFromSQL(dataSource, "SELECT * from PERSON");
-    }
-
     @RequestMapping(path = "/api/report/{report_id}")
     public ResponseEntity<Resource> getReport(@PathVariable("report_id") String report_id) {
 
@@ -74,6 +65,6 @@ public class Controller {
 
     @RequestMapping("/api/derictory/{modelName}")
     public String getData(@PathVariable("modelName") String name) {
-        return modelRepository.GetAllByName(name);
+        return modelUtils.GetAllByName(name);
     }
 }
