+++ source/server/src/main/java/registry/entity/menu/MenuItemRepository.java
@@ -1,4 +1,4 @@
-package registry.menu;
+package registry.entity.menu;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
