+++ source/server/src/main/java/registry/entity/model/ModelItemEntity.java
@@ -1,4 +1,4 @@
-package registry.model;
+package registry.entity.model;
 
 import javax.persistence.*;
