+++ source/server/src/main/java/registry/entity/report/ReportItemRepository.java
@@ -1,10 +1,8 @@
-package registry.report;
+package registry.entity.report;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.QueryHints;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.QueryHint;
 import java.util.List;
 
 @Repository
