--- source/server/src/main/java/registry/model/ModelRepository.java
@@ -1,10 +1,0 @@
-package registry.model;
-
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ModelRepository {
-    String GetAllByName (String modelName);
-}
